backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
google-generativeai
Flask-CORS 